requirements.txt
pycaw
python-dotenv
comtypes
pystray
Pillow
